Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.fillRect.transform-expected.txt

2d.fillRect.transform

fillRect is affected by transforms


PASS fillRect is affected by transforms
